Simplify query building in GetFeatures QR
- Group all select clause related fragments into one "selectClause" fragment
- Group all filter related fragments into existing filterWhereClause and simplify overriding it partially
- Pull up dataset identification column definition
- Pull up the orderBy fragment from IterateFeatures QR into GetFeatures QR
- Push down offset fragments into IterateFeatures QR
- Push down all tweaks related methods from GetFeaturesByBBox to GetFeaturesByBBoxTweaked
