+++ xyz-hub-service/src/main/java/com/here/xyz/httpconnector/config/query/ExportSpace.java
@@ -26,7 +26,7 @@ import java.sql.SQLException;
 
 //TODO: Remove that hack after refactoring is complete
 public interface ExportSpace<E extends SearchForFeaturesEvent> {
-  public SQLQuery buildQuery(E event) throws SQLException, ErrorResponseException;
+  SQLQuery buildQuery(E event) throws SQLException, ErrorResponseException;
 
   ExportSpace<E> withSelectionOverride(SQLQuery selectionOverride);
 
@@ -34,21 +34,18 @@ public interface ExportSpace<E extends SearchForFeaturesEvent> {
 
   ExportSpace<E> withCustomWhereClause(SQLQuery whereClauseOverride);
 
-  default SQLQuery patchQuery(SQLQuery query, SQLQuery geoOverride, SQLQuery customWhereClause, SQLQuery selectionOverride) {
-    if (geoOverride != null)
-      query.setQueryFragment("geo", geoOverride);
-    if (customWhereClause != null)
-      patchWhereClause(query, customWhereClause);
+  default SQLQuery patchSelectClause(SQLQuery selectClause, SQLQuery selectionOverride) {
     if (selectionOverride != null)
-      query.setQueryFragment("selection", selectionOverride);
-    return query.withQueryFragment("limit", "");
+      return selectClause.withQueryFragment("selection", selectionOverride);
+    return selectClause;
   }
 
-  private static void patchWhereClause(SQLQuery query, SQLQuery customWhereClause) {
-    SQLQuery filterWhereClause = query.getQueryFragment("filterWhereClause");
+  default SQLQuery patchWhereClause(SQLQuery filterWhereClause, SQLQuery customWhereClause) {
+    if (customWhereClause == null)
+      return filterWhereClause;
     SQLQuery customizedWhereClause = new SQLQuery("${{innerFilterWhereClause}} ${{customWhereClause}}")
         .withQueryFragment("innerFilterWhereClause", filterWhereClause)
         .withQueryFragment("customWhereClause", customWhereClause);
-    query.setQueryFragment("filterWhereClause", customizedWhereClause);
+    return customizedWhereClause;
   }
 }
