+++ xyz-psql-connector/src/main/resources/xyz_ext.sql
@@ -3109,48 +3109,6 @@ BEGIN
 END$$;
 ------------------------------------------------
 ------------------------------------------------
-CREATE OR REPLACE FUNCTION transform_load_features_input(ids TEXT[], versions BIGINT[])
-    RETURNS LOAD_FEATURE_VERSION_INPUT[] AS
-$BODY$
-DECLARE
-    output LOAD_FEATURE_VERSION_INPUT[];
-    item LOAD_FEATURE_VERSION_INPUT;
-BEGIN
-    IF coalesce(array_length(ids, 1),0) > 0 THEN
-        FOR i IN 1 .. array_upper(ids, 1)
-            LOOP
-                item := ROW(ids[i], versions[i]);
-                SELECT array_append(output, item) into output;
-
-            END LOOP;
-    END IF;
-    RETURN output;
-END
-$BODY$
-    LANGUAGE plpgsql VOLATILE;
-------------------------------------------------
-------------------------------------------------
-CREATE OR REPLACE FUNCTION transform_load_features_input(ids TEXT[])
-    RETURNS LOAD_FEATURE_VERSION_INPUT[] AS
-$BODY$
-DECLARE
-    output LOAD_FEATURE_VERSION_INPUT[];
-    versions BIGINT[];
-BEGIN
-    IF array_length(ids, 1) IS NULL THEN
-        RETURN output;
-    ELSE
-        FOR i IN 1 .. array_upper(ids, 1)
-        LOOP
-            versions := array_append(versions, max_bigint());
-        END LOOP;
-        RETURN transform_load_features_input(ids, versions);
-    END IF;
-END
-$BODY$
-LANGUAGE plpgsql VOLATILE;
-------------------------------------------------
-------------------------------------------------
 CREATE OR REPLACE FUNCTION xyz_create_history_partition(schema TEXT, rootTable TEXT, partitionNo BIGINT, partitionSize BIGINT)
     RETURNS VOID AS
 $BODY$
